requirements.txt
pandas
numpy
flask
pyyaml
scikit-learn
mlflow
dagshub
-e .